pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.2",
    "flask-httpauth>=4.8.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "python-decouple>=3.8",
    "python-dotenv>=1.2.1",
]
